refactor: FSD Phase 2a — UI キットを shared/ui へ移す (FRESTYLE-157)
FSD 移行(FRESTYLE-154)の Phase 2a。純粋な移動 + import 追従で振る舞いは変えない。

移動（すべて git mv により rename として記録）:
- components/ui      → shared/ui
- components/inkwell → shared/ui/inkwell
- components/icons   → shared/ui/icons

いずれもビジネス用語を含まない汎用部品。ビジネスを知っている部品
(LanguageBadge / CourseProgressBar / MessageBubble 等)は shared に置かず、
後続フェーズで entities / features へ振り分ける。

import は @/shared/ui 系の絶対パスへ機械的に置換。サブパス付き参照
(../components/ui/PageIntro 等)を最初の置換で取りこぼしたが、tsc の列挙で
検出して修正した。

Public API は既に名前付き re-export になっており FSD 規約に準拠（ワイルドカード
export は使っていない）。

検証: tsc / eslint / vitest(166ファイル1355件・assert 変更なし) / build /
E2E(12件) すべて緑。
