+++ frontend/src/pages/HelpPage.tsx
@@ -9,7 +9,7 @@ import {
   SparklesIcon,
   BookmarkIcon,
 } from '@heroicons/react/24/outline';
-import { PageIntro, StepIndicator, GuidedHint, GlossaryTerm, ActionCard } from '../components/ui';
+import { PageIntro, StepIndicator, GuidedHint, GlossaryTerm, ActionCard } from '@/shared/ui';
 import { GLOSSARY } from '../constants/glossary';
 
 /**
