+++ frontend/src/shared/ui/FilterChip.tsx
