+++ frontend/src/shared/ui/FirstTimeWelcome.tsx
