+++ frontend/src/shared/ui/GlossaryTerm.tsx
