+++ frontend/src/shared/ui/GuidedHint.tsx
