+++ frontend/src/shared/ui/HelpTooltip.tsx
