+++ frontend/src/shared/ui/PageIntro.tsx
