+++ frontend/src/shared/ui/StepIndicator.tsx
