+++ frontend/src/shared/ui/__tests__/FirstTimeWelcome.test.tsx
