+++ frontend/src/shared/ui/__tests__/GlossaryTerm.test.tsx
