+++ frontend/src/shared/ui/__tests__/GuidedHint.test.tsx
