+++ frontend/src/shared/ui/__tests__/StepIndicator.test.tsx
