+++ frontend/src/shared/ui/icons/FaviconIcon.tsx
