+++ frontend/src/shared/ui/inkwell/InkwellSkeleton.tsx
